// api/model/patchQuizzesQuestionsBulkBodyItem.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Live Quiz Arena API
 * Versioned backend contract for auth, user, quiz, and session
 * OpenAPI spec version: 1.0.0
 */

export type PatchQuizzesQuestionsBulkBodyItem = {
  id: string;
  questionText?: string;
  /** @minimum 0 */
  points?: number;
  /** @minimum 0 */
  timeLimit?: number;
};